# api/src/swagger.yaml
openapi: 3.0.3
info:
    title: FantasyPad
    description: Provided FantasyPad API endpoints.
    version: 1.0.0
paths:
...
    /api/auth/sign-in:
        post:
            summary: 'Returns an access token and refresh token(in cookie).'
            description: "**Important!** sign message must be look like: I'm owner of wallet with (user wallet). Authorization code: (user code)"
            tags:
                - api/auth
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SignInInput'
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                oneOf:
                                    - $ref: '#/components/responses/SignIn_200_SuccessResponse'
                                    - $ref: '#/components/responses/SignIn_200_FailResponse'
                    headers:
                        Set-Cookie:
                            $ref: '#/components/headers/SignIn-Set-Cookie'
                    description: ''
                '400':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/responses/BadRequest'
                '500':
                    description: '**Internal server error**'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/responses/InternalServerError'
    /api/auth/refresh:
        get:
            summary: 'Request new access token'
            tags:
                - api/auth
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                oneOf:
                                    - $ref: '#/components/responses/Refresh_200_SuccessResponse'
                                    - $ref: '#/components/responses/Refresh_200_FailResponse'
                    headers:
                        Set-Cookie:
                            $ref: '#/components/headers/Refresh-Set-Cookie'
                '500':
                    description: '**Internal server error**'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/responses/InternalServerError'
...
    /api/admin/auth/sign-in:
        post:
            summary: 'Returns an access token and refresh token(in cookie).'
            description: "**Important!** sign message must be look like: I'm owner of wallet with (admin wallet). Authorization code: (admin auth code)"
            tags:
                - api/admin/auth
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SignInAdminInput'
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                oneOf:
                                    - $ref: '#/components/responses/SignInAdmin_200_SuccessResponse'
                                    - $ref: '#/components/responses/SignInAdmin_200_FailResponse'
                    headers:
                        Set-Cookie:
                            $ref: '#/components/headers/SignIn-Set-Cookie'
                '400':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/responses/BadRequest'
                '500':
                    description: '**Internal server error**'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/responses/InternalServerError'
    /api/admin/auth/refresh:
        get:
            summary: 'Request new access token'
            tags:
                - api/admin/auth
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                oneOf:
                                    - $ref: '#/components/responses/RefreshAdmin_200_SuccessResponse'
                                    - $ref: '#/components/responses/RefreshAdmin_200_FailResponse'
                    headers:
                        Set-Cookie:
                            $ref: '#/components/headers/Refresh-Set-Cookie'
                '500':
                    description: '**Internal server error**'
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/responses/InternalServerError'
...
components:
    schemas:
...
        SignInInput:
            type: object
            properties:
                wallet:
                    type: string
                    description: Wallet address
                code:
                    type: string
                    description: uuid code from /api/auth/code
                signature:
                    type: string
                    description: "hex-like digest of message - I'm owner of wallet with (user wallet). Authorization code: (user code)"
        SignInAdminInput:
            type: object
            properties:
                wallet:
                    type: string
                    description: Wallet address
                code:
                    type: string
                    description: uuid code from /api/admin/auth/code
                signature:
                    type: string
                    description: "hex-like digest of message - I'm owner of wallet with (admin wallet). Authorization code: (admin auth code)"
...
    responses:
        BadRequest:
            type: object
            properties:
                error:
                    type: string
                    enum: ['Bad request']
                message:
                    type: array
                    items:
                        type: string
        InternalServerError:
            type: object
            properties:
                error:
                    type: string
                errorUniqId:
                    type: string
                message:
                    type: string
                stack:
                    type: string
            required:
                - error
                - errorUniqId
...
        SignInAdmin_200_SuccessResponse:
            type: object
            properties:
                success:
                    type: 'true'
                data:
                    type: string
                    description: JWT token
        SignIn_200_SuccessResponse:
            type: object
            properties:
                success:
                    type: 'true'
                data:
                    type: string
                    description: JWT token
        SignIn_200_FailResponse:
            type: object
            properties:
                success:
                    type: 'false'
                data:
                    type: object
                    properties:
                        error:
                            type: string
                            enum:
                                [
                                    'cannotFindAuthCode',
                                    'authCodeExpired',
                                    'signatureIsNotValid',
                                    'cannotFindUser',
                                ]
        SignInAdmin_200_FailResponse:
            type: object
            properties:
                success:
                    type: 'false'
                data:
                    type: object
                    properties:
                        error:
                            type: string
                            enum:
                                [
                                    'cannotFindAuthCode',
                                    'authCodeExpired',
                                    'signatureIsNotValid',
                                    'cannotFindAdmin',
                                ]
        Refresh_200_SuccessResponse:
            type: object
            properties:
                success:
                    type: 'true'
                data:
                    type: object
                    properties:
                        accessToken:
                            type: string
                            description: JWT token
        Refresh_200_FailResponse:
            type: object
            properties:
                success:
                    type: 'false'
                data:
                    type: object
                    properties:
                        error:
                            type: string
                            enum: ['cannotFindUser']
        RefreshAdmin_200_SuccessResponse:
            type: object
            properties:
                success:
                    type: 'true'
                data:
                    type: object
                    properties:
                        accessToken:
                            type: string
                            description: JWT token
        RefreshAdmin_200_FailResponse:
            type: object
            properties:
                success:
                    type: 'true'
                data:
                    type: object
                    properties:
                        error:
                            type: string
                            enum: ['cannotFindAdmin']
...
    headers:
        Refresh-Set-Cookie:
            description: 'Return when success is **true**'
            schema:
                type: string
                example: 'Refresh=*jwt_token*; HttpOnly; Path=/; Max-Age=1 day'
        SignIn-Set-Cookie:
            description: 'Return when success is **true**'
            schema:
                type: string
                example: 'Refresh=*jwt_token*; HttpOnly; Path=/; Max-Age=1 day'
...